requirements.txt
pillow
pip install psycopg2-binary
django-admin-interface